refactor testutil
