+++ .gitignore
@@ -15,3 +15,4 @@ __pycache__
 /tests/.vscode
 /tests/.idea
 /tests/MANIFEST
+*.DS_Store
